requirements_test.txt
nose==1.3.0
django-nose==1.1
